Added document_id in organisation
