+++ src/src/beneficiaries/beneficiaries.module.ts
@@ -20,6 +20,7 @@ import { HasuraModule as HasuraModuleFromServices } from '../services/hasura/has
 import { KeycloakModule } from '../services/keycloak/keycloak.module';
 import { BeneficiariesCoreService } from './beneficiaries.core.service';
 import { BeneficiariesService } from './beneficiaries.service';
+import { IpMiddleware } from 'src/common/middlewares/ip.middleware';
 @Module({
 	imports: [
 		UserModule,
@@ -74,7 +75,9 @@ export class BeneficiariesModule implements NestModule {
 
 			);
 		*/
-
+		consumer
+			.apply(IpMiddleware)
+			.forRoutes('/beneficiaries//admin/list', '/beneficiaries/:id');
 		consumer
 			.apply(CohortMiddleware)
 			.exclude(
