+++ src/src/common/middlewares/cohort.middleware.ts
@@ -12,22 +12,22 @@ export class CohortMiddleware implements NestMiddleware {
 	async use(req: any, res: Response, next: () => void) {
 		let goToNextMw = false;
 		//check IP User ID is present or not [x-ip-user-id]
-		if (
-			req?.headers?.['x-ip-org-id'] ||
-			req?.mw_roles?.includes('program_owner')
-		) {
-			req.mw_ip_user_id = req.headers['x-ip-org-id'];
+		// if (
+		// 	req?.headers?.['x-ip-org-id'] ||
+		// 	req?.mw_roles?.includes('program_owner')
+		// ) {
+		// 	req.mw_ip_user_id = req.headers['x-ip-org-id'];
 
-			const ip_user_id = parseInt(req.mw_ip_user_id, 10);
-			if (isNaN(ip_user_id)) {
-				return res.json({
-					success: false,
-					message: `${
-						req.mw_ip_user_id ? 'Invalid' : 'Required'
-					} Ip org Id`,
-				});
-			}
-		}
+		// 	const ip_user_id = parseInt(req.mw_ip_user_id, 10);
+		// 	if (isNaN(ip_user_id)) {
+		// 		return res.json({
+		// 			success: false,
+		// 			message: `${
+		// 				req.mw_ip_user_id ? 'Invalid' : 'Required'
+		// 			} Ip org Id`,
+		// 		});
+		// 	}
+		// }
 
 		if (req?.headers?.['x-program-id']) {
 			req.mw_program_id = req.headers['x-program-id'];
