+++ src/src/organisation/organisation.service.ts
@@ -67,7 +67,9 @@ export class OrganisationService {
 			program_id: request.mw_program_id,
 			academic_year_id: request.mw_academic_year_id,
 			status: 'active',
-			document_id: body?.document_id,
+			doc_per_cohort_id: body?.doc_per_cohort_id,
+			doc_per_monthly_id: body?.doc_per_monthly_id,
+			doc_quarterly_id: body?.doc_quarterly_id,
 			// Other fields as needed
 		};
 
@@ -89,8 +91,8 @@ export class OrganisationService {
 	}
 
 	public async getOrganisation(body: any, req: any, resp: any) {
-		const academic_year_id = body?.academic_year_id;
-		const program_id = body?.program_id;
+		const academic_year_id = req.mw_academic_year_id;
+		const program_id = req.mw_program_id;
 
 		try {
 			const page = isNaN(body.page) ? 1 : parseInt(body.page);
@@ -215,7 +217,9 @@ export class OrganisationService {
               program_id
               academic_year_id
               status
-							document_id
+							doc_per_cohort_id
+							doc_per_monthly_id
+							doc_quarterly_id
             }
           }
         }
